feat: add Hydrogen advanced caching example site (#32)
* feat: add Hydrogen example site

more or less as-is from https://github.com/netlify/hydrogen-template
with just some README updates and a `netlify.toml` tweak for the
monorepo

* perf: load cart, customer client-side to make pages cacheable

See
https://shopify.dev/docs/storefronts/headless/hydrogen/cart/render-client-side.

* feat: add full page caching

Cache most pages for 1 hour on the Netlify CDN with stale-while-revalidate (allowing 1 min for
background revalidation to complete). Cache these also on browsers (and anything else upstream) but
require a conditional request to revalidate.

On the search page, vary on the search input.

Disable this caching on data loaders that have personalization: account/*, cart.

* feat: add default caching of data resource routes

* feat: on-demand tag-based cache revalidation

* docs(readme): describe advanced caching features

This is a separate commit so I can link to it from the guide. :)